+++ src/FSharpx.Extras/CSharpCompat.fs
@@ -5,8 +5,9 @@ open System.IO
 open System.Net
 open System.Collections.Generic
 open System.Runtime.CompilerServices
-open FSharpx.Control.StreamReaderExtensions
 open FSharpx.Collections
+open FSharpx.Functional
+open FSharpx.Control
 open FSharpx