+++ tests/FSharpx.Tests/LensTests.fs
@@ -5,7 +5,8 @@ open NUnit.Framework
 open FsCheck
 open FsCheck.NUnit
 open FSharpx
-open FSharpx.Lens.Operators
+open FSharpx.Functional
+open FSharpx.Functional.Lens.Operators
 open FSharpx.Text
 
 type Car = {
@@ -102,7 +103,7 @@ let stateMonad() =
     Assert.AreEqual(tom.Salary, r)
     Assert.AreEqual(1100, tom1.Salary)
 
-open FSharpx.Lens.StateOperators
+open FSharpx.Functional.Lens.StateOperators
 
 [<Test>]
 let stateMonadOperators() =